Move uavcan bridge to a separate module
