+++ goldorak_base/src/uavcan_bridge.h
@@ -1,0 +1,8 @@
+#ifndef UAVCAN_BRIDGE_H
+#define UAVCAN_BRIDGE_H
+
+#include <stdlib.h>
+
+void uavcan_bridge_start(int id);
+
+#endif /* UAVCAN_BRIDGE_H */
